add logout and fix dialog ok click

+++ app/src/main/java/com/malec/cheesetime/ui/main/cheese/cheeseList/CheeseListFragment.kt
@@ -18,7 +18,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.malec.cheesetime.R
 import com.malec.cheesetime.di.Injectable
 import com.malec.cheesetime.model.CheeseSort
-import com.malec.cheesetime.ui.main.DeleteDialog
+import com.malec.cheesetime.ui.main.AlertDialogBuilder
 import com.malec.cheesetime.util.DateTimePicker
 import kotlinx.android.synthetic.main.fragment_cheese_list.*
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -80,7 +80,7 @@ class CheeseListFragment : Fragment(), Injectable {
                 true
             }
             R.id.deleteButton -> {
-                DeleteDialog(requireContext()).setOnOkButtonClickListener {
+                AlertDialogBuilder(requireContext()).setOnOkButtonClickListener {
                     viewModel.deleteSelected()
                 }.showCheeseDialog(viewModel.selectedCount.value)
