+++ app/src/main/java/com/malec/cheesetime/ui/main/cheese/cheeseManage/CheeseManageActivity.kt
@@ -29,7 +29,7 @@ import com.malec.cheesetime.model.Cheese
 import com.malec.cheesetime.model.Photo
 import com.malec.cheesetime.ui.BaseActivity
 import com.malec.cheesetime.ui.Screens
-import com.malec.cheesetime.ui.main.DeleteDialog
+import com.malec.cheesetime.ui.main.AlertDialogBuilder
 import com.malec.cheesetime.ui.main.ResultNavigator
 import com.malec.cheesetime.ui.main.cheese.cheeseManage.CheeseManageViewModel.Companion.CAMERA
 import com.malec.cheesetime.util.DateFormatter
@@ -88,7 +88,7 @@ class CheeseManageActivity : BaseActivity() {
         when (item.itemId) {
             R.id.saveButton -> saveCheese()
             R.id.deleteButton ->
-                DeleteDialog(this).setOnOkButtonClickListener {
+                AlertDialogBuilder(this).setOnOkButtonClickListener {
                     viewModel.deleteCheese()
                 }.showCheeseDialog()
             android.R.id.home -> onBackPressed()
